chore: cargo fmt + cargo clippy (0 warnings)
- Auto-fix: 19 sugestões aplicadas via clippy --fix
- Manual: struct literal Default, matches! macro, sort_by_key,
  remove &* deref, remove mut desnecessário
- Allow clippy::large_enum_variant no TabType

+++ src/ui/fund/panel/profit.rs
@@ -193,8 +193,7 @@ impl ProfitUI {
                         self.send_profit_message(cnpj, start_date, now);
                     }
                     FilterMonth::Ytd => {
-                        let start_date =
-                            chrono::NaiveDate::from_ymd_opt(now.year() as i32, 1, 1).unwrap();
+                        let start_date = chrono::NaiveDate::from_ymd_opt(now.year(), 1, 1).unwrap();
                         self.send_profit_message(cnpj, start_date, now);
                     }
                     FilterMonth::Custom => {}
